Conflicts:
	lib/gpumisc
	src/adapters/playback.cpp
	src/tfr/stft.cpp
	src/tools/playbackcontroller.cpp
	src/tools/rendercontroller.cpp
	src/ui/mainwindow.ui

+++ src/filters/ellipse.cpp
@@ -5,8 +5,8 @@
 #include "tfr/chunk.h"
 
 // gpumisc
-#include <TaskTimer.h>
-#include <computationkernel.h>
+#include "TaskTimer.h"
+#include "computationkernel.h"
 
 // std
 #include <iomanip>
@@ -50,9 +50,9 @@ bool Ellipse::
     TIME_FILTER TaskTimer tt("Ellipse");
 
     Area area = {
-            _centre_t * chunk.sample_rate - chunk.chunk_offset.asFloat(),
+            (float)(_centre_t * chunk.sample_rate - chunk.chunk_offset.asFloat()),
             chunk.freqAxis.getFrequencyScalarNotClamped( _centre_f ),
-            _centre_plus_radius_t * chunk.sample_rate - chunk.chunk_offset.asFloat(),
+            (float)(_centre_plus_radius_t * chunk.sample_rate - chunk.chunk_offset.asFloat()),
             chunk.freqAxis.getFrequencyScalarNotClamped( _centre_plus_radius_f )};
 
     ::removeDisc( chunk.transform_data,
